+++ code/src/sixsq/nuvla/server/resources/nuvlabox.clj
@@ -954,6 +954,7 @@ particular NuvlaBox release.
         components (:components body)]
     (nuvlabox-log/create-log nuvlabox components opts)))
 
+
 (defmethod crud/do-action [resource-type "create-log"]
   [{{uuid :uuid} :params :as request}]
   (try
@@ -961,7 +962,29 @@ particular NuvlaBox release.
       (-> (db/retrieve id request)
         (a/throw-cannot-manage request)
         (depl-utils/throw-can-not-do-action utils/can-create-log? "create-log")
-        (create-log request)))
+        (create-log request)))))
+
+;;
+;; Allows for the creation of a NuvlaBox API key on-demand
+;;
+
+(defn create-new-api-key
+  [{:keys [id state] :as nuvlabox}]
+  (if (#{state-decommissioned state-decommissioning} state)
+    (logu/log-and-throw-400 (str "invalid state for generating new API key: " state))
+    (do
+      (log/warn "generating new API key for nuvlabox:" id)
+      (let [[_ credential] (wf-utils/create-nuvlabox-api-key nuvlabox "")]
+        (r/json-response credential)))))
+
+
+(defmethod crud/do-action [resource-type "generate-new-api-key"]
+  [{{uuid :uuid} :params :as request}]
+  (try
+    (let [id (str resource-type "/" uuid)]
+      (-> (db/retrieve id request)
+          (a/throw-cannot-manage request)
+          (create-new-api-key)))
     (catch Exception e
       (or (ex-data e) (throw e)))))
 
@@ -999,6 +1022,7 @@ particular NuvlaBox release.
         disable-host-mgmt-op (u/action-map id :disable-host-level-management)
         enable-emergency-op  (u/action-map id :enable-emergency-playbooks)
         create-log-op        (u/action-map id :create-log)
+        generate-new-key-op  (u/action-map id :generate-new-api-key)
         ops                  (cond-> []
                                      (a/can-edit? resource request) (conj edit-op)
                                      (and (a/can-delete? resource request)
@@ -1041,7 +1065,10 @@ particular NuvlaBox release.
                                           (#{state-activated
                                              state-commissioned
                                              state-decommissioning
-                                             state-error} state)) (conj create-log-op))]
+                                             state-error} state)) (conj create-log-op)
+                                     (and (a/can-manage? resource request)
+                                          (not= state state-decommissioned)
+                                          (not= state state-decommissioning)) (conj generate-new-key-op))]
     (assoc resource :operations ops)))
 
 ;;