+++ code/src/sixsq/nuvla/server/resources/nuvlabox_log.clj
@@ -5,8 +5,6 @@ These resources represent the logs of a nuvlabox.
   (:require
     [sixsq.nuvla.auth.acl-resource :as a]
     [sixsq.nuvla.auth.utils :as auth]
-    [clojure.pprint :refer [pprint]]
-
     [sixsq.nuvla.db.filter.parser :as parser]
     [sixsq.nuvla.server.resources.common.crud :as crud]
     [sixsq.nuvla.server.resources.common.std-crud :as std-crud]