+++ code/test/sixsq/nuvla/server/resources/nuvlabox_2_lifecycle_test.clj
@@ -1455,6 +1455,103 @@
               (ltu/is-key-value :host-level-management-api-key nil)))))))
 
 
+(deftest create-activate-generate-new-key-lifecycle
+  (binding [config-nuvla/*stripe-api-key* nil]
+    (let [session       (-> (ltu/ring-app)
+                            session
+                            (content-type "application/json"))
+
+          session-owner (header session authn-info-header "user/alpha user/alpha group/nuvla-user group/nuvla-anon")]
+
+      #_{:clj-kondo/ignore [:redundant-let]}
+      (let [nuvlabox-id  (-> session-owner
+                             (request base-uri
+                                      :request-method :post
+                                      :body (json/write-str valid-nuvlabox))
+                             (ltu/body->edn)
+                             (ltu/is-status 201)
+                             (ltu/location))
+
+            nuvlabox-url (str p/service-context nuvlabox-id)
+
+            activate-url (-> session-owner
+                             (request nuvlabox-url)
+                             (ltu/body->edn)
+                             (ltu/is-status 200)
+                             (ltu/is-operation-present :generate-new-api-key)
+                             (ltu/get-op-url :activate))]
+
+        ;; activate nuvlabox
+        (-> session
+            (request activate-url
+                     :request-method :post)
+            (ltu/body->edn)
+            (ltu/is-status 200))
+
+        (let [session-nuvlabox    (header session authn-info-header
+                                          (str nuvlabox-id " " nuvlabox-id
+                                               " group/nuvla-nuvlabox group/nuvla-anon"))
+
+              generate-new-key    (-> session-nuvlabox
+                                      (request nuvlabox-url)
+                                      (ltu/body->edn)
+                                      (ltu/is-status 200)
+                                      (ltu/get-op-url :generate-new-api-key))
+
+              credential-url      (-> session-owner
+                                      (request generate-new-key
+                                               :request-method :post)
+                                      (ltu/body->edn)
+                                      (ltu/is-status 200)
+                                      (ltu/is-key-value (comp not str/blank?) :secret-key true)
+                                      (ltu/body)
+                                      :api-key
+                                      (ltu/href->url))
+
+              credential-nuvlabox (-> session-owner
+                                      (request credential-url)
+                                      (ltu/body->edn)
+                                      (ltu/is-status 200)
+                                      (ltu/is-key-value :parent nuvlabox-id)
+                                      (ltu/body))
+
+              claims              (:claims credential-nuvlabox)
+
+              decommission-url    (-> session-owner
+                                      (request nuvlabox-url)
+                                      (ltu/body->edn)
+                                      (ltu/is-status 200)
+                                      (ltu/get-op-url :decommission))]
+
+          ;; check ACL and claims of generated credential.
+          (is (= (:identity claims) nuvlabox-id))
+          (is (= (-> claims :roles set) #{nuvlabox-id
+                                          "group/nuvla-user"
+                                          "group/nuvla-anon"
+                                          "group/nuvla-nuvlabox"}))
+
+          ;; checks created api credential for NB visible for owner
+          (-> session-owner
+              (request credential-url)
+              (ltu/body->edn)
+              (ltu/is-status 200)
+              (ltu/is-operation-absent :edit)
+              (ltu/is-operation-absent :delete))
+
+          ;; decommission
+          (-> session-owner
+              (request decommission-url)
+              (ltu/body->edn)
+              (ltu/is-status 202))
+
+          ;; op is no longer present
+          (-> session-owner
+              (request nuvlabox-url)
+              (ltu/body->edn)
+              (ltu/is-status 200)
+              (ltu/is-operation-absent :generate-new-api-key)))))))
+
+
 (deftest create-activate-commission-get-context-lifecycle
   (binding [config-nuvla/*stripe-api-key* nil]
     (let [session       (-> (ltu/ring-app)